requirements.txt
attrs==19.1.0
requests==2.22.0
seaborn==0.9.0
pandas==0.25.1
matplotlib==3.1.1
